+++ packages/blade/src/components/Table/types.ts
@@ -75,6 +75,11 @@ type TableProps<Item> = {
    */
   data: TableData<Item>;
   /**
+   * Selection mode determines how the table rows can be selected.
+   * @default 'row'
+   **/
+  multiSelectTrigger?: 'checkbox' | 'row';
+  /**
    * The selectionType prop determines the type of selection that is allowed on the table.
    * The selectionType prop can be 'none', 'single' or 'multiple'.
    * @default 'none'
@@ -84,7 +89,25 @@ type TableProps<Item> = {
    * The onSelectionChange prop is a function that is called when the selection changes.
    * The function is called with an object that has a values property that is an array of the selected rows.
    **/
-  onSelectionChange?: ({ values }: { values: TableNode<Item>[] }) => void;
+  onSelectionChange?: ({
+    values,
+    selectedIds,
+  }: {
+    /**
+     * Note: on server side paginated data, this prop will only contain the selected rows on the current page.
+     *
+     * Thus, it's recommended to use `selectedIds` for more consistent state management across server/client paginated data.
+     *
+     * *Deprecated:* Use `selectedIds` instead.
+     *
+     * @deprecated
+     */
+    values: TableNode<Item>[];
+    /**
+     * An array of selected row ids.
+     */
+    selectedIds: Identifier[];
+  }) => void;
   /**
    * The isHeaderSticky prop determines whether the table header is sticky or not.
    * The default value is `false`.